+++ src/main/java/com/example/schedule_application/user/service/UserService.java
@@ -1,9 +1,13 @@
 package com.example.schedule_application.user.service;
 
-import com.example.schedule_application.user.dto.CreateUserRequest;
+import com.example.schedule_application.schedule.exception.NoPermissionException;
+import com.example.schedule_application.user.dto.LoginRequest;
+import com.example.schedule_application.user.dto.SignUpRequest;
 import com.example.schedule_application.user.dto.UpdateUserRequest;
 import com.example.schedule_application.user.dto.UserAllDetailsResponse;
 import com.example.schedule_application.user.entity.User;
+import com.example.schedule_application.user.exception.DuplicateEmailException;
+import com.example.schedule_application.user.exception.PasswordNotMatchException;
 import com.example.schedule_application.user.exception.UserNotFoundException;
 import com.example.schedule_application.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +15,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -18,15 +23,42 @@ public class UserService {
 
     private final UserRepository userRepository;
 
-    private User userValidation(Long userId) {
-        return userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
+    @Transactional
+    public User userValidation(Long userId) {
+        return userRepository.findById(userId).orElseThrow(
+                () -> new UserNotFoundException("존재하지 않는 유저입니다.")
+        );
+    }
+
+    public void validateOwner(Long userId, Long sessionUserId) {
+        if(!Objects.equals(userId, sessionUserId)) {
+            throw new NoPermissionException();
+        }
     }
 
     @Transactional
-    public UserAllDetailsResponse signUp(CreateUserRequest request) {
-        User user = new User(request.getName(), request.getEmail());
+    public UserAllDetailsResponse signUp(SignUpRequest request) {
+        if(userRepository.existsByEmail(request.getEmail()))
+            throw new DuplicateEmailException();
+
+        User user = new User(request.getName(), request.getEmail(), request.getPassword());
         User savedUser = userRepository.save(user);
         return UserAllDetailsResponse.from(savedUser);
+    }
+
+    @Transactional(readOnly = true)
+    public User login(LoginRequest request) {
+        User user = userRepository.findByEmail(request.email());
+
+        if(Objects.isNull(user)) {
+            throw new UserNotFoundException();
+        }
+
+        if(!user.getPassword().equals(request.password())) {
+            throw new PasswordNotMatchException();
+        }
+
+        return user;
     }
 
     @Transactional(readOnly = true)
@@ -43,15 +75,17 @@ public class UserService {
         return UserAllDetailsResponse.from(findedUser);
     }
 
-
-    public UserAllDetailsResponse update(Long userId, UpdateUserRequest request) {
+    @Transactional
+    public UserAllDetailsResponse update(Long userId, UpdateUserRequest request, Long sessionUserId) {
+        validateOwner(userId, sessionUserId);
         User user = userValidation(userId);
         user.updateUserDetails(request.getName(), request.getEmail());
         return UserAllDetailsResponse.from(user);
     }
 
     @Transactional
-    public void delete(Long userId) {
+    public void delete(Long userId, Long sessionUserId) {
+        validateOwner(userId, sessionUserId);
         userValidation(userId);
         userRepository.deleteById(userId);
     }
